Updated roll call
Roll call now opens as a modal from anywhere. Roll call page is now removed. Updated routes
