+++ client/routes.js
@@ -10,7 +10,6 @@ import Interactions from "./components/interactions/Interactions.js";
 import RegisteredUsers from "./components/users-list/RegisteredUsers";
 import LoggedInUsers from "./components/users-list/LoggedInUsers";
 import MainSettingsPage from "./components/settings/MainSettingsPage";
-import RollCall from "./components/roll-call/RollCall";
 import Message from './components/ContentMessage/Message';
 import Login from "./components/authentication/login/Login";
 import Register from "./components/authentication/register/Register";
@@ -52,10 +51,6 @@ const routes = [
             {
                 path: '/settings',
                 element: <MainSettingsPage/>,
-            },
-            {
-                path: '/roll-call',
-                element: <RollCall/>
             },
             {
                 path: '/ContentMessage',
